* slight logic fix * fixed visual sheathing of players (seeing other with weapons out)
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2848 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/MiscHandler.cpp
@@ -1466,7 +1466,7 @@ void WorldSession::HandleSetSheathedOpcode( WorldPacket & recv_data )
 {
 	uint32 active;
 	recv_data >> active;
-	_player->SetByte(UNIT_FIELD_BYTES_2,0,active); 
+	_player->SetByte(UNIT_FIELD_BYTES_2,0,(uint8)active); 
 }
 
 void WorldSession::HandlePlayedTimeOpcode( WorldPacket & recv_data )
